refactor: Move an integration-style test to the `cli` module
Interoperability between package managers should be tested on a higher
level now that package managers become independent plugins.
