+++ cli/src/funTest/assets/projects/synthetic/spdx-subproject-conan/project-xyz.spdx.yml
