+++ cli/src/funTest/assets/projects/synthetic/spdx-subproject-conan/subproject/conanfile.py
