+++ cli/src/funTest/assets/projects/synthetic/spdx-subproject-conan/subproject/subproject-with-conan-reference.spdx.yml
